requirements.txt
allure_python_commons
pytest
selenium
